chore: move utility functions to common crate (#615)
* chore: move utility functions to common crate

* fix: cleanup

* chore: move get previous proof

* chore: cleanup dependencies
