+++ trace_decoder/Cargo.toml
@@ -50,7 +50,6 @@ criterion = { workspace = true }
 glob = "0.3.1"
 libtest-mimic = "0.7.3"
 plonky2_maybe_rayon = { workspace = true }
-pretty_env_logger = { workspace = true }
 prover = { workspace = true }
 serde_json = { workspace = true }
 serde_path_to_error = { workspace = true }
