+++ zero_bin/common/src/env.rs
@@ -1,0 +1,16 @@
+use std::io;
+
+use dotenvy::dotenv;
+use tracing::warn;
+
+/// Attempt to load in the local `.env` if present and set any environment
+/// variables specified inside of it.
+///
+/// To keep things simple, any IO error we will treat as the file not existing
+/// and continue moving on without the `env` variables set.
+pub fn load_dotenvy_vars_if_present() {
+    match dotenv() {
+        Ok(_) | Err(dotenvy::Error::Io(io::Error { .. })) => (),
+        Err(e) => warn!("Found local `.env` file but was unable to parse it! (err: {e})",),
+    }
+}
