+++ zero_bin/leader/Cargo.toml
@@ -10,20 +10,18 @@ categories.workspace = true
 build = "../common/build.rs"
 
 [dependencies]
-paladin-core = { workspace = true }
-clap = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
+alloy = { workspace = true }
 anyhow = { workspace = true }
+axum = { workspace = true }
+clap = { workspace = true }
+futures = { workspace = true }
+paladin-core = { workspace = true }
 serde = { workspace = true }
-dotenvy = { workspace = true }
-tokio = { workspace = true }
 serde_json = { workspace = true }
 serde_path_to_error = { workspace = true }
-futures = { workspace = true }
-alloy.workspace = true
-axum = { workspace = true }
-toml = { workspace = true }
+tokio = { workspace = true }
+tracing = { workspace = true }
+tracing-subscriber = { workspace = true }
 
 # Local dependencies
 evm_arithmetization = { workspace = true }
