+++ zero_bin/rpc/Cargo.toml
@@ -16,6 +16,7 @@ anyhow = { workspace = true }
 clap = { workspace = true }
 futures = { workspace = true }
 hex = { workspace = true }
+itertools = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
@@ -23,7 +24,6 @@ tower = { workspace = true, features = ["retry"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 url = { workspace = true }
-itertools = { workspace = true }
 
 # Local dependencies
 compat = { workspace = true }
@@ -34,9 +34,9 @@ trace_decoder = { workspace = true }
 zero_bin_common = { workspace = true }
 
 [build-dependencies]
+anyhow = { workspace = true }
 cargo_metadata = { workspace = true }
 vergen = { workspace = true }
-anyhow = { workspace = true }
 
 [features]
 default = ["eth_mainnet"]
