--- zero_bin/worker/src/init.rs
@@ -1,11 +1,0 @@
-use tracing_subscriber::{prelude::*, util::SubscriberInitExt, EnvFilter};
-pub(crate) fn tracing() {
-    tracing_subscriber::Registry::default()
-        .with(
-            tracing_subscriber::fmt::layer()
-                .with_ansi(false)
-                .compact()
-                .with_filter(EnvFilter::from_default_env()),
-        )
-        .init();
-}
